requirements.txt
fastapi
uvicorn
pymysql
python-dotenv
bcrypt
pyjwt
